+++ R/fit2df.R
@@ -20,7 +20,7 @@
 #' @param exp Currently GLM only. Exponentiate coefficients and confidence
 #'   intervals. Defaults to TRUE.
 #' @param confint_type One of \code{c("profile", "default")} for GLM models
-#'   (\code{\link[MASS]{confint.glm}}) or \code{c("profile", "Wald", "boot")}
+#'   (\link[MASS:confint]{confint.glm}) or \code{c("profile", "Wald", "boot")}
 #'   for \code{glmer/lmer} models (\code{\link[lme4]{confint.merMod}}.). Not
 #'   implemented for \code{lm, coxph or coxphlist}.
 #' @param confint_level The confidence level required.
@@ -284,6 +284,7 @@ fit2df.glmboot = function(.data, condense=TRUE, metrics=FALSE, remove_intercept=
 													estimate_suffix = "",
 													p_name = "p",
 													digits=c(2,2,3),
+													exp = TRUE,
 													confint_sep = "-", ...){
 	if(metrics == TRUE) warning("Metrics not currently available for this model")
 	
@@ -296,19 +297,21 @@ fit2df.glmboot = function(.data, condense=TRUE, metrics=FALSE, remove_intercept=
 	
 	df.out = data.frame(
 		explanatory = names(x$t0),
-		estimate = exp(x$t0))
+		estimate = x$t0)
 	for (i in 1:dim(df.out)[1]){
-		df.out$L95[i] = exp(sort(x$t[,i]))[floor(R*0.025)]
-		df.out$U95[i] = exp(sort(x$t[,i]))[floor((R*0.975)+1)]
+		df.out$L95[i] = sort(x$t[,i])[floor(R*0.025)]
+		df.out$U95[i] = sort(x$t[,i])[floor((R*0.975)+1)]
 		df.out$p[i] = ifelse(x$t0[i] >= 0, mean(x$t[,i]<0)*2, mean(x$t[,i]>0)*2)
 	}
-	df.out$estimate = round(df.out$estimate, d.estimate)
-	df.out$L95 = round(df.out$L95, d.confint)
-	df.out$U95 = round(df.out$U95, d.confint)
-	df.out$p = round(df.out$p, d.p)
+	
+	if(exp){
+		df.out = df.out %>% 
+			dplyr::mutate_at(dplyr::vars(estimate, L95, U95), ~ exp(.))
+	}
+	
 	colnames(df.out) = c(explanatory_name, paste0(estimate_name, estimate_suffix), "L95", "U95", p_name)
 	
-	if (condense==TRUE){
+	if(condense){
 		df.out = condense_fit(df.out, explanatory_name=explanatory_name,
 													estimate_name=estimate_name, estimate_suffix=estimate_suffix,
 													p_name=p_name, digits=digits, confint_sep=confint_sep)
@@ -619,6 +622,44 @@ fit2df.crr <- function(.data, condense=TRUE, metrics=FALSE,
 		return(df.out)
 	}
 }
+
+
+#' Extract \code{coxme::coxme} model fit results to dataframe: \code{finalfit} model extracters
+#'
+#' \code{fit2df.coxme} is the model extract method for \code{eoxme::\link[coxme]{coxme}}.
+#'
+#' @rdname fit2df
+#' @method fit2df coxme
+#' @export
+#'
+fit2df.coxme <- function(.data, condense=TRUE, metrics=FALSE,
+											 explanatory_name = "explanatory",
+											 estimate_name = "HR",
+											 estimate_suffix = "",
+											 p_name = "p",
+											 digits=c(2,2,3),
+											 confint_sep = "-", ...){
+	
+	df.out = extract_fit(.data=.data, explanatory_name=explanatory_name,
+											 estimate_name=estimate_name, estimate_suffix=estimate_suffix,
+											 p_name=p_name)
+	
+	if (condense==TRUE){
+		df.out = condense_fit(.data=df.out, explanatory_name=explanatory_name,
+													estimate_name=estimate_name, estimate_suffix=estimate_suffix,
+													p_name=p_name, digits=digits, confint_sep=confint_sep)
+	}
+	# Extract model metrics
+	if (metrics==TRUE){
+		metrics.out = ff_metrics(.data)
+		return(list(df.out, metrics.out))
+	} else {
+		return(df.out)
+	}
+}
+
+
+
 
 #' Extract \code{crruni} and \code{crrmulti} model fit results to dataframe:
 #' \code{finalfit} model extracters